+++ packages/terra-form/src/Field.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-
+import IconError from 'terra-icon/lib/icon/IconError';
 import 'terra-base/lib/baseStyles';
 import styles from './Field.scss';
 
@@ -13,17 +13,29 @@ const propTypes = {
    */
   children: PropTypes.node,
   /**
-   * Error message for when the input is invalid.
+   * Error message for when the input is invalid on validation.
    */
   error: PropTypes.node,
+  /**
+   * Error Icon to display when the input is invalid on validation.
+   */
+  errorIcon: PropTypes.element,
   /**
    * Help element to display with the input.
    */
   help: PropTypes.node,
   /**
+   * Determines whether or not to hide the required indicator on the label.
+   */
+  hideRequired: PropTypes.bool,
+  /**
    * The htmlFor attribute on the field label.
    */
   htmlFor: PropTypes.string,
+  /**
+   * Determines whether the field is in error or not.
+   */
+  isInvalid: PropTypes.bool,
   /**
    * Determines whether the field is an inline field.
    */
@@ -40,24 +52,56 @@ const propTypes = {
    * Determines whether the field is required.
    */
   required: PropTypes.bool,
+  /**
+   * Determines whether or not to append the 'optional' label to a non-required field label.
+   */
+  showOptional: PropTypes.bool,
 };
 
 const defaultProps = {
   children: null,
   error: null,
+  errorIcon: <IconError />,
   help: null,
+  hideRequired: false,
   htmlFor: undefined,
+  isInvalid: false,
   isInline: false,
   label: null,
   labelAttrs: {},
   required: false,
+  showOptional: false,
 };
 
-const Field = ({ children, error, help, htmlFor, isInline, label, labelAttrs, required, ...customProps }) => {
+const contextTypes = {
+  /* eslint-disable consistent-return */
+  intl: (context) => {
+    if (context.intl === undefined) {
+      return new Error('Please add locale prop to Base component to load translations');
+    }
+  },
+};
+
+const Field = (props, { intl }) => {
+  const {
+    children,
+    error,
+    errorIcon,
+    help,
+    hideRequired,
+    htmlFor,
+    isInvalid,
+    isInline,
+    label,
+    labelAttrs,
+    required,
+    showOptional,
+    ...customProps
+  } = props;
+
   const fieldClasses = cx([
     'field',
     { 'field-inline': isInline },
-    { 'field-required': required },
     customProps.className,
   ]);
 
@@ -66,17 +110,34 @@ const Field = ({ children, error, help, htmlFor, isInline, label, labelAttrs, re
     labelAttrs.className,
   ]);
 
+  let labelGroup;
+  if (label) {
+    labelGroup = (
+      <div className={cx('label-group')}>
+        {isInvalid && <div className={cx('error-icon')}>{errorIcon}</div>}
+        {<label htmlFor={htmlFor} {...labelAttrs} className={labelClassNames}>
+          {required && (isInvalid || !hideRequired) && <div className={cx('required')}>*</div>}
+          {label}
+          {required && !isInvalid && hideRequired && <div className={cx('required-hidden')}>*</div>}
+          {showOptional && !required && <div className={cx('optional')}>{intl.formatMessage({ id: 'Terra.form.field.optional' })}</div>}
+        </label>}
+        {!isInvalid && <div className={cx('error-icon-hidden')}>{errorIcon}</div>}
+      </div>
+    );
+  }
+
   return (
     <div {...customProps} className={fieldClasses}>
-      {label && <label htmlFor={htmlFor} {...labelAttrs} className={labelClassNames}>{label}</label>}
+      {labelGroup}
       {children}
-      {help && <small className={cx('help-text')} tabIndex="-1">{help}</small>}
-      {error && <small className={cx('error')} tabIndex="-1">{error}</small>}
+      {isInvalid && error && <div className={cx('error-text')}>{error}</div>}
+      {help && <div className={cx('help-text')}>{help}</div>}
     </div>
   );
 };
 
 Field.propTypes = propTypes;
 Field.defaultProps = defaultProps;
+Field.contextTypes = contextTypes;
 
 export default Field;