+++ packages/terra-form/tests/jest/Field.test.jsx
@@ -1,16 +1,81 @@
-/* eslint-disable react/jsx-boolean-value */
-
+/* eslint-disable react/jsx-boolean-value, import/no-extraneous-dependencies */
 import React from 'react';
+import IconHelp from 'terra-icon/lib/icon/IconHelp';
+import intlContexts from './intl-context-setup';
 import Field from '../../src/Field';
 
 it('should render a default component', () => {
-  const field = (<Field />);
-  const wrapper = shallow(field);
+  const field = <Field />;
+  const wrapper = shallow(field, intlContexts.shallowContext);
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render a Field when all the possible props are passed into it', () => {
-  const input = (
+it('should render a field label', () => {
+  const field = (
+    <Field
+      label="Field Label"
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an optional field label', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      showOptional
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a required field label', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      required
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a required feild label with required hidden', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      required
+      hideRequired
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a feild error message', () => {
+  const field = (
+    <Field
+      error="Error Text"
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a field help message', () => {
+  const field = (
+    <Field
+      help="Help Text"
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an inline field with most of the possible props are passed into it', () => {
+  const field = (
     <Field
       error="This field is required"
       help="This is a test input"
@@ -18,12 +83,58 @@ it('should render a Field when all the possible props are passed into it', () =>
       label="Text"
       labelAttrs={{ className: 'healtheintent-legend' }}
       isInline
-      required
     >
-      <input id="test" type="text" value="Test" />
+      <input id="test" type="text" />
     </Field>
   );
 
-  const wrapper = shallow(input);
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a field in error', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      error="Error Text"
+      help="Help Text"
+      isInvalid
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render an optional field in error', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      error="Error Text"
+      help="Help Text"
+      showOptional
+      isInvalid
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a required field in error', () => {
+  const field = (
+    <Field
+      label="Field Label"
+      error="Error Text"
+      help="Help Text"
+      required
+      isInvalid
+    />
+  );
+  const wrapper = shallow(field, intlContexts.shallowContext);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a field with a custom error icon', () => {
+  const field = <Field errorIcon={<IconHelp />} />;
+  const wrapper = shallow(field, intlContexts.shallowContext);
   expect(wrapper).toMatchSnapshot();
 });