+++ packages/terra-form/tests/jest/TextField.test.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import intlContexts from './intl-context-setup';
 import TextField from '../../src/TextField';
 
 it('should render a default TextField component', () => {
   const input = (<TextField />);
-  const wrapper = shallow(input);
+  const wrapper = shallow(input, intlContexts.shallowContext);
   expect(wrapper).toMatchSnapshot();
 });
 
@@ -24,18 +25,18 @@ it('should render a TextField with the rest of the props', () => {
       required
     />);
 
-  const wrapper = shallow(input);
+  const wrapper = shallow(input, intlContexts.shallowContext);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render as uncontrolled when just a default value is passed into the TextField', () => {
   const input = <TextField defaultValue="foo" />;
-  const wrapper = mount(input);
+  const wrapper = mount(input, intlContexts.mountContext);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render as controlled when just a default value is passed into the TextField', () => {
   const input = <TextField value="foo" onChange={() => {}} />;
-  const wrapper = mount(input);
+  const wrapper = mount(input, intlContexts.mountContext);
   expect(wrapper).toMatchSnapshot();
 });