+++ packages/terra-form/tests/nightwatch/number-field/DefaultNumberField.jsx
@@ -1,7 +1,13 @@
 import React from 'react';
-
+import Base from 'terra-base';
 import NumberField from '../../../lib/NumberField';
 
-const numberField = () => <NumberField id="default-number-field" />;
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
+
+const numberField = () => (
+  <Base locale={locale}>
+    <NumberField id="default-number-field" />
+  </Base>
+);
 
 export default numberField;