+++ packages/terra-form/tests/nightwatch/text-field/DefaultTextField.jsx
@@ -1,7 +1,13 @@
 import React from 'react';
-
+import Base from 'terra-base';
 import TextField from '../../../lib/TextField';
 
-const textField = () => <TextField id="default-text-field" />;
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
+
+const textField = () => (
+  <Base locale={locale}>
+    <TextField id="default-text-field" />
+  </Base>
+);
 
 export default textField;