+++ packages/terra-site/src/examples/form/examples/NumberField.jsx
@@ -5,8 +5,13 @@ import styles from '../../../site.scss';
 class NumberFieldExamples extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { input: '0.075' };
+    this.state = { input: '0.075', isInvalid: false };
     this.handleChange = this.handleChange.bind(this);
+    this.toggleIsInvalid = this.toggleIsInvalid.bind(this);
+  }
+
+  toggleIsInvalid() {
+    this.setState({ isInvalid: !this.state.isInvalid });
   }
 
   handleChange(event) {
@@ -30,10 +35,13 @@ class NumberFieldExamples extends React.Component {
             inputAttrs={{ className: 'healtheintent-application' }}
             inputId="tax-rate"
             isInline
+            isInvalid={this.state.isInvalid}
             onChange={this.handleChange}
           />
         </form>
-        <hr /><p>Input Provided: <span className={styles['site-input-display']}>{this.state.input}</span></p>
+        <hr />
+        <p>If a number feild is invalid, an error icon will be displayed. <button onClick={this.toggleIsInvalid}>Toggle Invalid State</button></p>
+        <p>Input Provided: <span className={styles['site-input-display']}>{this.state.input}</span></p>
         <br />
       </div>
     );