+++ packages/terra-site/src/examples/form/examples/SelectField.jsx
@@ -5,8 +5,13 @@ import styles from '../../../site.scss';
 class SelectFieldExamples extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { selected: 'snappers' };
+    this.state = { selected: 'snappers', isInvalid: false };
     this.handleChange = this.handleChange.bind(this);
+    this.toggleIsInvalid = this.toggleIsInvalid.bind(this);
+  }
+
+  toggleIsInvalid() {
+    this.setState({ isInvalid: !this.state.isInvalid });
   }
 
   handleChange(event) {
@@ -25,14 +30,18 @@ class SelectFieldExamples extends React.Component {
                         { value: 'joeys', display: 'Joeys' },
                         { value: 'micros', display: 'Microprocessors' }]}
             help="TO DETERMINE IF YOU ARE A HUMAN, PLEASE PICK YOUR FAVORITE SMALL ANIMAL"
+            error="This Field is Required"
             label="Human Test"
             name="zibby"
             defaultValue="snappers"
             required
+            isInvalid={this.state.isInvalid}
             onChange={this.handleChange}
           />
         </form>
-        <hr /><p>Option Selected: <span className={styles['site-input-display']}>{this.state.selected}</span></p>
+        <hr />
+        <p>If a select feild is invalid, an error icon will be displayed. <button onClick={this.toggleIsInvalid}>Toggle Invalid State</button></p>
+        <p>Option Selected: <span className={styles['site-input-display']}>{this.state.selected}</span></p>
         <br />
       </div>
     );